Fixed Beautiful NBT showing values inside short and byte arrays as doubles

+++ src/main/java/net/pedroricardo/commander/BeautifulNbt.java
@@ -44,14 +44,14 @@ public class BeautifulNbt {
         boolean comma = false;
         if (o instanceof byte[]) {
             result.append("B; ");
-            for (double element : ((byte[])o)) {
+            for (byte element : ((byte[])o)) {
                 if (comma) result.append(", ");
                 result.append(element(element));
                 comma = true;
             }
         } else if (o instanceof short[]) {
             result.append("S; ");
-            for (double element : ((short[])o)) {
+            for (short element : ((short[])o)) {
                 if (comma) result.append(", ");
                 result.append(element(element));
                 comma = true;
